+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 1.0b4 (unreleased)
 ------------------
 
+- Fix content types displayed on navigation (https://colab.interlegis.leg.br/ticket/2961).
+  [hvelarde]
+
 - Add ods, odt, odp, html, csv, zip, tgz, ppt, pptx, xls and xlsx to the list
   of file types that can be uploaded to the portal (https://colab.interlegis.leg.br/ticket/2966).
   [hvelarde]